requirements.txt
addict=2.0.0
beautifulsoup4==4.4.0
pyyaml==3.11
requests==2.7.0
